refactor and addressing comment
