+++ runtime/window_data.cc
@@ -1,0 +1,12 @@
+// Copyright 2013 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+#include "flutter/runtime/window_data.h"
+
+namespace flutter {
+WindowData::WindowData() = default;
+
+WindowData::~WindowData() = default;
+
+}  // namespace flutter
