+++ shell/platform/darwin/ios/framework/Source/FlutterDartProject_Internal.h
@@ -6,7 +6,7 @@
 #define SHELL_PLATFORM_IOS_FRAMEWORK_SOURCE_FLUTTERDARTPROJECT_INTERNAL_H_
 
 #include "flutter/common/settings.h"
-#include "flutter/runtime/runtime_controller.h"
+#include "flutter/runtime/window_data.h"
 #include "flutter/shell/common/engine.h"
 #include "flutter/shell/platform/darwin/ios/framework/Headers/FlutterDartProject.h"
 
@@ -15,7 +15,7 @@ NS_ASSUME_NONNULL_BEGIN
 @interface FlutterDartProject ()
 
 - (const flutter::Settings&)settings;
-- (const flutter::WindowData&)windowData;
+- (const flutter::WindowData)defaultWindowData;
 
 - (flutter::RunConfiguration)runConfiguration;
 - (flutter::RunConfiguration)runConfigurationForEntrypoint:(nullable NSString*)entrypointOrNil;
